[build_machinery] Single module signing support
Summary:
Allow signing to fail.
Previously, it would just swallow the error, as there's nothing bad in a background shell dying.
Now failure creates a marker file, which is then used to fail the build.

There's a couple more places where signing is done, but there's no need to do this trick there. bzl/kernel.bzl and bzl/klp.bzl don't have this rpm building magic, they just have straight invocations of autograph, which should fail properly.

This diff is against an upstream branch that I used for testing, but I'll put this to 6.11, 6.13, and build machinery.

Test Plan:
` sudo ./.buck2 build --no-buckd --local-only -c build_info.provided=True -c build_info.kernelversion=6.12.41 -c build_info.gittish=1_g88bc2c23c0b6 -c build_info.disable_kmod= -c build_info.fbk_tag=v6.12.41-fbk1 -c build_info.use_aarch64_emul=0  --show-full-json-output //:uname-nvidia-open-kmod-570.124.06-1-6.12.41-0_fbk1_cvmhost //:kernelversion -c kernel.module=nvidia-open-kmod-570.124.06-1 -c kernel.version=6.12.41-0_fbk1_cvmhost -c kernel.arch=x86_64 - c build_info.custom_tag= -c kernel.sign_mod=true -c kernel.sign_mod_key=fbinfra-kmod-hsm-WRONG -c kernel.devel_rpm=/data/sandcastle/temp/skycastle/tmpdv5by61_/kernel-devel-6.12.41-0_fbk1_cvmhost_1_g88bc2c23c0b6.x86_64.rpm //:standalone_module`
pay attention to `kernel.sign_mod_key=fbinfra-kmod-hsm-WRONG` there's no such key, and now it failed and I got the log of the build container on the screen, it was apparent that there's no such key from the output.

Positive test in D78124439 in this stack.

Rollback Plan:

Reviewers: riel, gustavold, #kernel_pe, davidhwei, leit

Reviewed By: riel

Subscribers: ps, scottbpc, jkkm

Differential Revision: https://phabricator.intern.facebook.com/D80140172

Tasks: T136749058

Tags: oncall, oncallteam-kernel-infra, aarch64, oncallteam-kernel-monitoring

+++ facebook/build/modules.bzl
@@ -134,6 +134,7 @@ def module(name, arch, module, kernel_devel, uname=None, dependencies=None, loca
 %{__os_install_post}\\\\
 touch /rw/BUILDROOT/do-signature\\\\
 while [ -e /rw/BUILDROOT/do-signature ]; do sleep 10; done\\\\
+
 %{nil}
 EOF
 """,
@@ -148,7 +149,7 @@ EOF
         native.genrule(
             name = name + "-sign-wrapper",
             cmd = """
-                cat > "$OUT" <<EOF
+                cat > "$OUT" <<'SIGNEOF'
 #!/bin/sh
 
 # Start the signature waiter in a background subshell
@@ -164,10 +165,34 @@ EOF
          ! -e "$(location :{name}-shared-build-dir)" ]; then
         exit
     fi
-    sudo autograph_client.par kmod --sign-key {sign_key} --kernel-tree "$(location :{name}-shared-build-dir)"
-    rm "$(location :{name}-shared-build-dir)/do-signature"
+
+    KERNEL_TREE=$(location :{name}-shared-build-dir)
+
+    # /failed-signature is a signal file to the build process to signify that
+    # the signature failed. If it exists then the build process will fail.
+    # Sign modules via SandcastleKernelBuildCommand (Autograph-whitelisted)
+    SIGN_TMPDIR=\$(mktemp -d)
+    SIGN_LOG=$SIGN_TMPDIR/signing.log
+    (
+        set -xe
+        tar -cf $SIGN_TMPDIR/unsigned_modules.tar -C "$KERNEL_TREE" .
+        EVERSTORE_HANDLE=\$(clowder put $SIGN_TMPDIR/unsigned_modules.tar --fbtype EVERSTORE_LINUX_KERNEL)
+        SCUTIL_OUTPUT=\$(scutil create '{{"command":"SandcastleKernelBuildCommand","args":{{"build_mode":"signing","everstore_handle":"'$EVERSTORE_HANDLE'","sign_modules_key":"{sign_key}"}},"capabilities":{{"type":"lego","vcs":"linux-kernel-git","tenant":"kernel"}},"nonce":"'$SANDCASTLE_NONCE'","alias":"kernel-module-signing","oncall":"kernel_infra","hash":"master"}}' --await --follow-retries -v json 2>&1)
+        INSTANCE_ID=\$(echo "$SCUTIL_OUTPUT" | jq -r '.id')
+        STEP_LOG_HANDLE=\$(scutil get-log-info $INSTANCE_ID | grep 'handle:' | tail -1 | awk '{{print $2}}')
+        SIGNED_HANDLE=\$(clowder get $STEP_LOG_HANDLE | grep -oP 'Signed Everstore handle: \K\S+')
+        clowder get $SIGNED_HANDLE > $SIGN_TMPDIR/signed_modules.tar
+        tar -xf $SIGN_TMPDIR/signed_modules.tar -C "$KERNEL_TREE"
+    ) > "$SIGN_LOG" 2>&1 || touch "$KERNEL_TREE/failed-signature"
+
+    echo "=== Signing log ===" >&2
+    cat "$SIGN_LOG" >&2
+    echo "=== End signing log ===" >&2
+    rm -rf "$SIGN_TMPDIR"
+
+    rm "$KERNEL_TREE/do-signature"
 ) &
-EOF
+SIGNEOF
                 chmod +x "$OUT"
 """.format(name = name, sign_key = sign_key),
             out = "signature-wrapper.sh",
@@ -220,6 +245,10 @@ EOF
             --define "kernel_version {uname}" \
             --define "using_clang 1" \
             --with llvm_cross || touch /rw/BUILDROOT/abort-signature
+            if [ -f /rw/BUILDROOT/failed-signature ]; then
+                echo "Failed to sign module"
+                exit 1
+            fi
             cp -R /root/rpmbuild/RPMS/{arch}/*.rpm /rw/rpms
         """.format(uname = "\$(cat /tmp/uname)", build_prep = build_prep, arch = arch),
         bind_ro = bind_ro,
@@ -631,4 +660,3 @@ modules = [
         archs = ["x86_64"],
     )
 ]
-
